+++ LibiadaMusicTests/BorodaDivider/FmotivIdentificatorTests.cs
@@ -47,22 +47,22 @@ namespace LibiadaMusicTests.BorodaDivider
 
             var fmid = new FmotivIdentifier();
             Assert.AreEqual(0,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[0].Id);
             Assert.AreEqual(1,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[1].Id);
             Assert.AreEqual(1,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[2].Id);
             Assert.AreEqual(1,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[3].Id);
             Assert.AreEqual(0,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[4].Id);
             Assert.AreEqual(0,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[5].Id);
 
         }
@@ -108,22 +108,22 @@ namespace LibiadaMusicTests.BorodaDivider
             var fmid = new FmotivIdentifier();
             //        FmotivChain fmidentedchain = new FmotivChain();
             Assert.AreEqual(0,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[0].Id);
             Assert.AreEqual(0,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[1].Id);
             Assert.AreEqual(0,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[2].Id);
             Assert.AreEqual(1,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[3].Id);
             Assert.AreEqual(2,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[4].Id);
             Assert.AreEqual(3,
-                fmid.GetIdentification(fmchain1, (int) ParamPauseTreatment.Ignore, (int) ParamEqualFM.Sequent)
+                fmid.GetIdentification(fmchain1, ParamPauseTreatment.Ignore, ParamEqualFM.Sequent)
                     .FmotivList[5].Id);
         }
     }
